Add a 1px offset to sideline textures
This fixes texturing issues I noticed on Windows x64 builds of Minetest. If the white line touches the edge of
the texture, the game will render a small white line on the opposite side of the texture as well. Shaders would
make this line even bigger. We should propably open an issue at the Minetest repository.